Struct win_kernel_driver::DriverBuilder

pub struct DriverBuilder { /* fields omitted */ }

Use this to build a kernel driver object you can interact with

Example

let driver_bin = include_bytes!("../winRing0.sys");
let driver = DriverBuilder::new()
             .set_device_description("winRing0 driver")
             .set_device_id("WinRing0_1_2_0")
             .set_driver_bin(driver_bin.to_vec())
             .build().unwrap();

Methods

impl DriverBuilder

pub fn new() -> Self

pub fn set_device_id(self, device_id: &'static str) -> Self

Set the device id (required)

pub fn set_device_description(self, device_description: &'static str) -> Self

Set the device description (required)

pub fn set_device_type(self, device_type: DWORD) -> Self

Set the device type (defaults to FILE_DEVICE_UNKNOWN (0x00000022))

pub fn set_driver_path(self, driver_path: PathBuf) -> Self

Set the path to the driver file (optional)

pub fn set_driver_bin(self, driver_bin: Vec<u8>) -> Self

Use a bytearray for the driver. It will be written to a temporary location Useful with the !include_bin macro

pub fn build(&mut self) -> Result<WinKernelDriver, String>

Build a WinKernelDriver instance